pyproject.toml
[project]
name = "openclaw-sdk"
version = "0.1.0"
description = "Python SDK for OpenClaw Gateway"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.7",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openclaw"]

[project.urls]
Homepage = "https://github.com/6xdax/openclaw-ws-client"
Repository = "https://github.com/6xdax/openclaw-ws-client"
